+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixes
 
+- Restore compatibility with WebMocks Typhoeus instrumentation
+- Catch WebMock exception and bubble them to the request promises
+
 ### Breaks
 
 ## 2.0.0 - (2025-02-14)